+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 # [Unreleased]
 
+# 2.0.0-rc.4
+
 - Include source maps for Parchment
 - **Clipboard** Support pasting links copied from iOS share sheets
 - Fix config parsing where undefined values were kept